refactor(tools): migrate web tools to directory structure
- Refactored HTTP, Scraper, HTML Parser, and URL Validator tools
- Split monolithic files into organized directory structures:
  - http-client.ts (130 lines) -> http/ directory (3 tools)
  - scraper.ts (151 lines) -> scraper/ directory (1 tool)
  - html-parser.ts (158 lines) -> html-parser/ directory (3 tools)
  - url-validator.ts (154 lines) -> url-validator/ directory (3 tools)
- Created types.ts files with schemas and interfaces for each tool group
- Split 10 tools total into individual files in tools/ subdirectories
- Created index.ts files with default exports and factory functions
- Updated web/index.ts to export from new locations
- All 975 tests passing
- Build successful
- No breaking changes to public API

Follows the pattern established by Confluence, web-search, and Slack tools.
Part of Phase 3 of the tools refactoring plan.
