+++ packages/tools/src/web/index.ts
@@ -5,9 +5,9 @@
  */
 
 export * from './http/index.js';
-export * from './scraper.js';
-export * from './html-parser.js';
-export * from './url-validator.js';
+export * from './scraper/index.js';
+export * from './html-parser/index.js';
+export * from './url-validator/index.js';
 export * from './web-search/index.js';
 export * from './slack/index.js';
 export * from './confluence/index.js';
